Add principal curvature feature
I'm adding a principal curvature function, as well as all the project infrastructure (CMake, vcpkg, etc.). This is a big commit and essentially a rewriting of history, because my git history became very convoluted as I figured things out. Now that I know what I'm doing (future me while find the naivety of this statement funny), I reverted back to the initial commit and re-added all the code for a cleaner commit history without the hundreds the trial-and-error commits.